ai_agents_hospitality-api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=12.0
jinja2==3.1.6
pydantic-settings>=2.0.0
python-multipart>=0.0.6
pyyaml>=6.0.0

# LangChain dependencies for Exercise 0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-google-genai>=1.0.0
langchain-openai>=0.1.0  # For OpenAI-compatible proxy support
langchain-community>=0.2.0



